+++ src/main/java/com/edee/foundationsforfaith/records/DonationRecord.java
@@ -1,6 +1,6 @@
 package com.edee.foundationsforfaith.records;
 
-public record DonationRecord(String donorMessage, Integer donationAmount, String projectName, String stoneEmail) {
+public record DonationRecord(String donorMessage, Float donationAmount, String projectName, String stoneEmail) {
     public String thanks() {
         return "Thanks for your donation of " + donationAmount + " to " + projectName + ".";
     }
